Refactor conformance golden scenario into modular pipeline
